refactor: remove unused Query and StreamQuery functions from OpenAIAgent to streamline code
